requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-google-genai
pypdf
python-dotenv
Faker
tiktoken